requirements.txt
absl-py==2.1.0
addict==2.4.0
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
anndata==0.10.9
antlr4-python3-runtime==4.9.3
anyio==4.6.0
array_api_compat==1.8
arrow==1.3.0
asttokens
async-lru==2.0.4
async-timeout==4.0.3
asyncio-atexit==1.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
biothings_client==0.4.1
blessed==1.20.0
boto3==1.35.25
botocore==1.35.25
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm
contourpy==1.3.0
crcmod==1.7
croniter==1.3.15
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.11
datasets==2.19.1
dateutils==0.6.12
debugpy
decorator
deepdiff==7.0.1
Deprecated==1.2.15
diffdist==0.1
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
editor==1.6.6
einops==0.8.0
exceptiongroup
executing
fastapi==0.88.0
filelock==3.14.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.2.3
future==1.0.0
get-annotations==0.1.2
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.66.0
grpcio==1.66.2
h11==0.14.0
h5py==3.11.0
harmonypy==0.0.10
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.25.0
hydra-core==1.3.2
hyperopt==0.1.2
idna==3.10
importlib_metadata
importlib_resources==6.4.5
inquirer==3.4.0
ipykernel
ipython
isodate==0.7.2
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jupyter_client
jupyter_core
kiwisolver==1.4.7
legacy-api-wrap==1.4
lightly==1.5.15
lightly-utils==0.0.2
lightning==1.9.1
lightning-cloud==0.5.70
lightning-utilities==0.11.7
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.9.2
matplotlib-inline
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.5
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mygene==3.2.2
natsort==8.4.0
nest_asyncio
networkx==3.2.1
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-httpx==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-instrumentation-threading==0.49b2
opentelemetry-instrumentation-urllib==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==1.5.0
parso
patsy==0.5.6
pexpect
pickleshare
pillow==10.4.0
platformdirs
prettytable==3.11.0
prompt_toolkit
protobuf==5.28.2
psutil==5.9.8
ptyprocess
pure_eval
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.18
Pygments
PyJWT==2.9.0
pymongo==4.10.1
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil
python-multipart==0.0.10
pytorch-lightning==1.9.0
pytz==2023.4
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq
rdflib==7.1.3
readchar==4.2.0
regex==2024.9.11
requests==2.28.2
rich==13.4.2
runs==1.2.2
s3transfer==0.10.2
safetensors==0.4.5
scanpy==1.10.3
scikit-learn==1.5.2
scikit-misc==0.3.1
scipy==1.13.1
scprep==1.2.3
scvi==0.6.8
seaborn==0.13.2
sentry-sdk==2.19.0
session_info==1.0.0
setproctitle==1.3.4
six
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data
starlette==0.22.0
starsessions==1.3.0
statsmodels==0.14.3
stdlib-list==0.10.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.19.1
tomli==2.0.2
torch==1.13.1+cu117
torch-geometric==2.6.0
torch-scatter==2.1.2
torchaudio==0.13.1+cu117
torchmetrics==1.4.2
torchtext==0.14.1
torchtoolbox==0.1.8.2
torchvision==0.14.1+cu117
tornado
tqdm==4.65.0
traitlets
transformers==4.44.0
triton==3.1.0
types-python-dateutil==2.9.0.20240906
typing_extensions
tzdata==2024.1
umap-learn==0.5.6
urllib3==1.26.20
uvicorn==0.30.6
wandb==0.18.7
wcwidth
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.16.0
xlrd==2.0.1
xmod==1.8.1
xxhash==3.5.0
yapf==0.40.2
yarl==1.11.1
zipp==3.20.2